+++ Order-Core/src/main/java/com/goormgb/be/ordercore/match/dto/response/MatchListByDateResponse.java
@@ -5,9 +5,9 @@ import java.time.LocalDate;
 import java.util.List;
 
 import com.goormgb.be.domain.club.entity.Club;
-import com.goormgb.be.domain.entity.Stadium;
 import com.goormgb.be.domain.match.entity.Match;
 import com.goormgb.be.domain.match.enums.SaleStatus;
+import com.goormgb.be.domain.statium.entity.Stadium;
 
 public record MatchListByDateResponse(
 	LocalDate date,
