+++ common-core/src/main/java/com/goormgb/be/domain/match/entity/Match.java
@@ -3,8 +3,8 @@ package com.goormgb.be.domain.match.entity;
 import java.time.Instant;
 
 import com.goormgb.be.domain.club.entity.Club;
-import com.goormgb.be.domain.entity.Stadium;
 import com.goormgb.be.domain.match.enums.SaleStatus;
+import com.goormgb.be.domain.statium.entity.Stadium;
 import com.goormgb.be.global.entity.BaseEntity;
 
 import jakarta.persistence.Column;
